Update Go to 1.24.6 (#31784) (#31824)
Cherry pick for https://github.com/fleetdm/fleet/pull/31784.

We decided to patch it because 1.24.6 contains a fix for a HIGH CVE:
[CVE-2025-47907](https://nvd.nist.gov/vuln/detail/CVE-2025-47907)
```
Cancelling a query (e.g. by cancelling the context passed to one of the query methods) during a call
to the Scan method of the returned Rows can result in unexpected results if other queries are being
made in parallel. This can result in a race condition that may overwrite the expected results with those
of another query, causing the call to Scan to return either unexpected results from the other
query or an error.
```