integrate new cert authorities API on frontend (#31696)
Relates to #30944

This PR integrates the new cert authorities APIs into the UI. There is
also a lot of cleanup to get rid of the code that was handling cert
authorities when it was on the config which is not needed anymore.

this PR includes:

- updating cert authority forms and their helpers to work with new cert
authority APIs
- updating and creating new cert authorities interfaces and types
- updating and creating new cert authorities entities and using them in
components
- removal of unneeded cert auth code in the app config
- small changes to ticket destination copy


- [x] Changes file added for user-visible changes in `changes/`,
`orbit/changes/` or `ee/fleetd-chrome/changes`.
- [x] Added/updated automated tests
- [x] QA'd all new/changed functionality manually

> NOTE: dummy data is being used on purpose and there will be another PR
to remove this when the backend API is complete