Touch software updated at timestamp if software inventory causes a no-op
Fir #31721. Unsure if this is the right fix because we no-op when last opened time hasn't changed by more than (by default) an hour, and touching the timestamp would remove any sign that data is potentially stale.

Alternate solution of "have the frontend use the details updated timestamp" doesn't feel great either though.

Also found a likely more straightforward repro of #28584 while writing tests for this, or at least what feels like a more straightforward repro of that bug, but fixing that bug is out of scope for fixing this one.