Change job title from 'Product Design Engineer' to 'Product Design' (#32325)
@noahtalerman I think we should revert this to "Product Designer" and
just be clear in the job requirements and interview process that it's a
heavy technical role.

Why? Because the responsibilities of this role are the same as Rachael
and Marko, who are both "Product Designer". This introduced a new
"Product Design Engineer" role that is functionally no different than
the existing "Product Designer" roles, so IMO it adds confusion.