Update entra-conditional-access-integration.md
Expanded group assignment for the conditional access policy. To help admins understand that if they want to properly enforce access for certain applications, they need to scope the broadest group possible for their application. This will make sure that any devices not managed by Fleet will be prompted to enroll in Fleet and be marked as compliant before access is granted.