fix issues with batch script summary (#32516)
for #32468 
for #32466

# Details

This PR fixes the following issues:

1. When viewing the summary for a batch script scheduled for the future,
clicking the "Pending" hosts number now shows the correct hosts targeted
by the batch, rather than an empty list.
2. When viewing the summary for a batch script that has been canceled,
the number of canceled hosts will correctly be displayed in the
"canceled" row rather than the "pending row".
3. When viewing the summary for any batch script, the table now splits
up Error and Incompatible states and clicking on either goes to the
correct list of hosts in that state.
4. When a batch script run has all of its targeted hosts deleted, it
will now be cleaned up and marked as "finished".

The first three issues are already fixed on the main branch, hence the
PR directly to the 4.73 release candidate. The last issue will be
patched on main in a separate PR (with tests) immediately following this
once.

# Checklist for submitter

If some of the following don't apply, delete the relevant line.

- [X] Input data is properly validated, `SELECT *` is avoided, SQL
injection is prevented (using placeholders for values in statements)

## Testing

- [ ] Added/updated automated tests
- [ ] Where appropriate, [automated tests simulate multiple hosts and
test for host
isolation](https://github.com/fleetdm/fleet/blob/main/docs/Contributing/reference/patterns-backend.md#unit-testing)
(updates to one hosts's records do not affect another)

- [X] QA'd all new/changed functionality manually
- [X] Started a new batch script to run "now" using osquery-perf, and
quit the agent halfway through the run so I could test the summary modal
with hosts in different states. Verified that the host list had the
correct # of hosts for each state.
- [X] Canceled that batch and verified that the # of hosts formerly
shown as "pending" in the summary modal now appeared as "canceled", and
verified that clicking that number showed the correct list of hosts.
- [X] Started a new batch script to run in the future, and verified that
clicking the "pending" hosts number showed the correct hosts in the list
- [X] Started a new batch script run with one host, deleted that host,
and triggered the `batch_activity_completion_checker` schedule, and
verified that the batch was moved to "finished".

For unreleased bug fixes in a release candidate, one of:

- [X] Confirmed that the fix is not expected to adversely impact load
test results