fix host cancelation reporting on canceled future batches (#32563)
for #32468

# Details

Previously, if you canceled a script that was scheduled for the future,
the hosts in that script were shown as "pending" in the batch summary
and in the filtered hosts list. This PR fixes the displays so that those
hosts are correctly shown as "canceled".

# Checklist for submitter

- [X] Input data is properly validated, `SELECT *` is avoided, SQL
injection is prevented (using placeholders for values in statements)

## Testing

- [X] Added/updated automated tests
- [X] Where appropriate, [automated tests simulate multiple hosts and
test for host
isolation](https://github.com/fleetdm/fleet/blob/main/docs/Contributing/reference/patterns-backend.md#unit-testing)
(updates to one hosts's records do not affect another)

- [X] QA'd all new/changed functionality manually

**Scheduled script**
1. Created a new batch script run scheduled for a future date
2. Canceled that batch run
3. Clicked on the batch run in the "finished" tab of the batch scripts
list
4. Verified that the number of canceled hosts = the number of targeted
hosts for that batch, and all other numbers were 0.
5. Verified that clicking the canceled hosts number navigated to the
correct list of canceled hosts

**"Run now" script**
1. Created a new batch script run with "run now"
2. Waited for at least one host to run.
3. Canceled that batch
4. Clicked on the batch run in the "finished" tab of the batch scripts
list
5. Verified that the number of canceled hosts = the number that were
still pending when I canceled the script, and that the # of pending
hosts was 0
5. Verified that clicking the canceled hosts number navigated to the
correct list of canceled hosts

**One more batch**
1. Created another batch script with the same hosts, scheduled for the
future
2. Verified that the "pending" host list is correct


For unreleased bug fixes in a release candidate, one of:

- [X] Confirmed that the fix is not expected to adversely impact load
test results