Fix formatting for clarity (#32658)
Trying to get lines 12 and 13 to appear on their own lines, instead of
as a continuation of the previous paragraph as show in the screen shot.
For bonus points, putting each value in step 11 on its own line for
clarity.

<img width="762" height="196" alt="Screenshot 2025-09-05 at 11 24 29 AM"
src="https://github.com/user-attachments/assets/dd357130-b996-480e-9a58-b594c30c5049"
/>

---------

Co-authored-by: Eric <eashaw@sailsjs.com>