User story template: Add "Usage statistics" checkbox (#33289)
Consider adding usage statistics as part of every user story. PRs
(wireframes) should be made to the JSON object in the usage statistics
guide:
https://fleetdm.com/guides/fleet-usage-statistics#what-is-included-in-usage-statistics-in-fleet



Why did this come up now? We're having to spend a lot of time to figure
out which customers to notify re breaking changes to the certificate
authorities (CAs) YAML:
https://docs.google.com/document/d/1w8_q4HyehCIoG1xDc-vD91xkBVajwbYbX_mwQWUtu9A/edit?tab=t.0