Allow host vitals docs edits to be approved by Secuirty & Compliance Tech Lead
Software ingestion via osquery is owned by S&C, and e.g. adding sources to software ingestion requires
revisions to detail queries. So, given that tech leads can approve these changes, might as well not
bottleneck on the orchestration tech lead for this, and the added reviewer gives a little bit more of a
heads-up when orchestration makes tweaks to detail queries for non-ingestion reasons (or when assisting
with ingestion-related work like we've been doing this sprint).