Add example YAML configuration for macOS setup
This came out of a customer call where we received the following output from the `fleetctl generate-gitops` command:
```
The macos_setup configuration is not supported by this tool yet.  To configure it, please follow the Fleet documentation at https://fleetdm.com/docs/configuration/yaml-files#macos-setup
```
In the moment, it wasn't clear to me or the customer what had to be done. This update provides the context (where the `macos_setup` configuration is supposed to exist) as well as an example of what it looks like.