Add host link on scheduled query report table (#34971)
<!-- Add the related story/sub-task/bug number, like Resolves #123, or
remove if NA -->
**Related issue:** Resolves #33249

# Details 

This PR makes the host names in the scheduled query report link to the
respective host query reports. The requirement of having the host names
in the live query report link to the host details page was already
fulfilled in #34019.

# Checklist for submitter

If some of the following don't apply, delete the relevant line.

## Testing

- [ ] Added/updated automated tests
trying but it's tricky to test a `Link` component's `href` without a
router.

- [X] QA'd all new/changed functionality manually

For unreleased bug fixes in a release candidate, one of:

- [X] Confirmed that the fix is not expected to adversely impact load
test results